sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentspan"
version = "0.1.0"
description = "Agentspan SDK — durable, scalable, observable AI agents"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
dependencies = [
    "conductor-python>=1.3.5-rc1",
    "httpx>=0.24",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.scripts]
agentspan = "agentspan.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
testing = [
    "litellm>=1.0",
]

[project.entry-points."pytest11"]
agentspan-testing = "agentspan.agents.testing.pytest_plugin"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring a live Agentspan server (deselect with '-m \"not integration\"')",
    "sse: marks tests requiring SSE streaming support (requires AGENTSPAN_STREAMING_ENABLED=true)",
    "agent_correctness: marks tests as agent correctness tests",
    "semantic: marks tests that use LLM judge for semantic assertions (requires litellm)",
]

[tool.coverage.run]
source = ["agentspan.agents"]

[tool.coverage.report]
show_missing = true
skip_empty = true
precision = 1

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
no_strict_optional = true
